fix: Player 객체와 PlayerCards 객체 분리
