+++ src/main/kotlin/blackjack/model/Player.kt
@@ -1,21 +1,14 @@
 package blackjack.model
 
-class Player(val name: String, private val deck: Deck) {
-    private val _cards: MutableList<Card> = mutableListOf()
-    val cards: List<Card>
-        get() = _cards.toList()
-
-    init {
-        _cards.addAll(deck.draw(2))
-    }
+class Player(val name: String, private val playerCards: PlayerCards) {
 
     fun addCard(isAdd: Boolean) {
         if (isAdd) {
-            _cards.addAll(deck.draw(1))
+            playerCards.add()
         }
     }
 
-    fun calculateCardScore(): Int {
-        return _cards.sumOf { it.cardNumber.score }
-    }
+    fun isBust(): Boolean = playerCards.calculateCardScore() > 21
+
+    fun getCards() = playerCards.cards
 }
